Refactor tests from workloads to procedures
