+++ benchmarks/osb/extensions/registry.py
@@ -4,8 +4,8 @@
 # this file be licensed under the Apache-2.0 license or a
 # compatible open source license.
 
-from helpers.param_sources import register as param_sources_register
-from helpers.runners import register as runners_register
+from .param_sources import register as param_sources_register
+from .runners import register as runners_register
 
 
 def register(registry):
