+++ benchmarks/osb/indices/faiss-index.json
@@ -2,8 +2,8 @@
   "settings": {
     "index": {
       "knn": true,
-      "number_of_shards": {{ target_index_primary_shards | default(1) }},
-      "number_of_replicas": {{ target_index_replica_shards | default(0) }}
+      "number_of_shards": {{ target_index_primary_shards }},
+      "number_of_replicas": {{ target_index_replica_shards }}
     }
   },
   "mappings": {
@@ -16,9 +16,9 @@
           "space_type": "{{ target_index_space_type }}",
           "engine": "faiss",
           "parameters": {
-            "ef_search":  {{ target_index_ef_search | default(512) }},
-            "ef_construction": {{ target_index_ef_construction | default(512) }},
-            "m": {{ target_index_m | default(16) }}
+            "ef_search":  {{ hnsw_ef_search }},
+            "ef_construction": {{ hnsw_ef_construction }},
+            "m": {{ hnsw_m }}
           }
         }
       }
