+++ benchmarks/osb/indices/model-index.json
@@ -8,9 +8,9 @@
   },
   "mappings": {
     "properties": {
-      "{{ target_field_name | default("target_field") }}": {
+      "{{ target_field_name }}": {
         "type": "knn_vector",
-        "model_id": "{{ model_id }}"
+        "model_id": "{{ train_model_id }}"
       }
     }
   }
