+++ benchmarks/osb/indices/nmslib-index.json
@@ -2,9 +2,9 @@
   "settings": {
     "index": {
       "knn": true,
-      "knn.algo_param.ef_search":  {{ target_index_ef_search | default(512) }},
-      "number_of_shards": {{ target_index_primary_shards | default(1) }},
-      "number_of_replicas": {{ target_index_replica_shards | default(0) }}
+      "knn.algo_param.ef_search":  {{ hnsw_ef_search }},
+      "number_of_shards": {{ target_index_primary_shards }},
+      "number_of_replicas": {{ target_index_replica_shards }}
     }
   },
   "mappings": {
@@ -17,8 +17,8 @@
           "space_type": "{{ target_index_space_type }}",
           "engine": "nmslib",
           "parameters": {
-            "ef_construction": {{ target_index_ef_construction | default(512) }},
-            "m": {{ target_index_m | default(16) }}
+            "ef_construction": {{ hnsw_ef_construction }},
+            "m": {{ hnsw_m }}
           }
         }
       }
