+++ benchmarks/osb/indices/train-index.json
@@ -1,0 +1,16 @@
+{
+  "settings": {
+    "index": {
+      "number_of_shards": {{ train_index_primary_shards }},
+      "number_of_replicas": {{ train_index_replica_shards }}
+    }
+  },
+  "mappings": {
+    "properties": {
+      "{{ train_field_name }}": {
+        "type": "knn_vector",
+        "dimension": {{ target_index_dimension }}
+      }
+    }
+  }
+}
