--- benchmarks/osb/knn-no-train/params.json
@@ -1,16 +1,0 @@
-{
-  "target_index_name": "target_index",
-  "target_index_body": "nmslib-index.json",
-  "target_field_name": "target_field",
-  "target_index_primary_shards": 3,
-  "target_index_replica_shards": 1,
-  "target_index_ef_search": 512,
-  "target_index_ef_construction": 512,
-  "target_index_m": 16,
-  "target_index_dimension": 128,
-  "target_index_space_type": "l2",
-  "bulk_size": 200,
-  "bulk_index_data_set_format": "hdf5",
-  "bulk_index_data_set_path": "/home/ec2-user/data/sift-128-euclidean.hdf5",
-  "bulk_index_clients": 10
-}
