--- benchmarks/osb/knn-no-train/workload.json
@@ -1,52 +1,0 @@
-{
-    "version": 2,
-    "description": "k-NN Plugin no train workload",
-    "indices": [
-        {
-            "name": "{{ target_index_name | default("target_index") }}",
-            "body": "{{ target_index_body | default("nmslib-index.json") }}"
-        }
-    ],
-    "schedule": [
-        {
-            "name": "delete-target-index",  
-            "index": "{{ target_index_name | default("target_index") }}",
-            "only-if-exists": true,
-            "operation": "delete-index"
-        },
-        {
-            "name": "create-target-index",
-            "index": "{{ target_index_name | default("target_index") }}",
-            "operation": "create-index"
-        },
-        {
-            "name": "wait-for-cluster-to-be-green",
-            "operation": "cluster-health",
-            "request-params": {
-                "wait_for_status": "green"
-            }
-        },
-        {
-            "operation": {
-                "name": "custom-vector-bulk",
-                "operation-type": "custom-vector-bulk",
-                "param-source": "bulk-from-data-set",
-                "index": "{{ target_index_name | default("target_index") }}",
-                "field": "{{ target_field_name | default("target_field") }}",
-                "bulk_size": {{ bulk_size | default(200) }},
-                "data_set_format": "{{ bulk_index_data_set_format }}",
-                "data_set_path": "{{ bulk_index_data_set_path }}"
-            },   
-            "clients": {{ bulk_index_clients | default(1) }}
-        },
-        {
-            "operation": {
-                "name": "refresh-target-index",
-                "operation-type": "custom-refresh",
-                "index": "{{ target_index_name|default("target_index") }}",
-                "index_name": "target_index",
-                "retries": 100
-            }
-        }
-    ]
-}
