--- benchmarks/osb/knn-train/params.json
@@ -1,29 +1,0 @@
-{
-  "target_index_name": "target_index",
-  "target_field_name": "target_field",
-  "target_index_primary_shards": 3,
-  "target_index_replica_shards": 1,
-  "target_index_dimension": 128,
-  "target_index_space_type": "l2",
-  "bulk_size": 200,
-  "bulk_index_data_set_format": "hdf5",
-  "bulk_index_data_set_path": "/home/ec2-user/data/sift-128-euclidean.hdf5",
-  "bulk_index_clients": 10,
-  "model_method": "ivfpq",
-  "model_id": "test-model",
-  "train_index_name": "train_index",
-  "train_field_name": "train_field",
-  "train_bulk_size": 200,
-  "train_index_data_set_format": "hdf5",
-  "train_index_data_set_path": "/home/ec2-user/data/sift-128-euclidean.hdf5",
-  "train_vector_limit": 1000000,
-  "train_bulk_index_clients": 10,
-  "train_index_primary_shards": 1,
-  "train_index_replica_shards": 0,
-  "train_search_size": 500,
-  "ivf_nlists": 10,
-  "ivf_nprobes": 1,
-  "pq_code_size": 8,
-  "pq_m": 8,
-  "train_timeout": 5000
-}
