--- benchmarks/osb/knn-train/train-index.json
@@ -1,16 +1,0 @@
-{
-  "settings": {
-    "index": {
-      "number_of_shards": {{ train_index_primary_shards | default(1) }},
-      "number_of_replicas": {{ train_index_replica_shards | default(0) }}
-    }
-  },
-  "mappings": {
-    "properties": {
-      "{{ train_field_name | default("train_field") }}": {
-        "type": "knn_vector",
-        "dimension": {{ target_index_dimension }}
-      }
-    }
-  }
-}
