+++ benchmarks/osb/operations/default.json
@@ -1,0 +1,53 @@
+[
+    {
+        "name": "ivfpq-train-model",
+        "operation-type": "train-model",
+        "model_id": "{{ model_id }}",
+        "timeout": {{ train_timeout }},
+        "body": {
+            "training_index": "{{ train_index_name }}",
+            "training_field": "{{ train_field_name }}",
+            "dimension": {{ target_index_dimension }},
+            "search_size": {{ train_search_size }},
+            "max_training_vector_count": {{ train_index_num_vectors }},
+            "method": {
+                "name":"ivf",
+                "engine":"faiss",
+                "space_type": "{{ target_index_space_type }}",
+                "parameters":{
+                    "nlist": {{ ivf_nlists }},
+                    "nprobes": {{ ivf_nprobes }},
+                    "encoder":{
+                        "name":"pq",
+                        "parameters":{
+                            "code_size": {{ pq_code_size }},
+                            "m": {{ pq_m }}
+                        }
+                    }
+                }
+            }
+        }
+    },
+    {
+        "name": "ivf-train-model",
+        "operation-type": "train-model",
+        "model_id": "{{ model_id }}",
+        "timeout": {{ train_timeout | default(1000) }},
+        "body": {
+            "training_index": "{{ train_index_name }}",
+            "training_field": "{{ train_field_name }}",
+            "search_size": {{ train_search_size }},
+            "dimension": {{ target_index_dimension }},
+            "max_training_vector_count": {{ train_index_num_vectors }},
+            "method": {
+                "name":"ivf",
+                "engine":"faiss",
+                "space_type": "{{ target_index_space_type }}",
+                "parameters":{
+                    "nlist": {{ ivf_nlists }},
+                    "nprobes": {{ ivf_nprobes }}
+                }
+            }
+        }
+    }
+]
