+++ benchmarks/osb/procedures/no-train-test.json
@@ -1,0 +1,48 @@
+{% import "benchmark.helpers" as benchmark with context %}
+{
+    "name": "no-train-test",
+    "default": true,
+    "schedule": [
+        {
+            "name": "delete-target-index",
+            "index": "{{ target_index_name }}",
+            "only-if-exists": true,
+            "operation": "delete-index"
+        },
+        {
+            "operation": {
+                "name": "create-target-index",
+                "operation-type": "create-index",
+                "index": "{{ target_index_name }}"
+            }
+        },
+        {
+            "name": "wait-for-cluster-to-be-green",
+            "operation": "cluster-health",
+            "request-params": {
+                "wait_for_status": "green"
+            }
+        },
+        {
+            "operation": {
+                "name": "custom-vector-bulk",
+                "operation-type": "custom-vector-bulk",
+                "param-source": "bulk-from-data-set",
+                "index": "{{ target_index_name }}",
+                "field": "{{ target_field_name }}",
+                "bulk_size": {{ target_index_bulk_size }},
+                "data_set_format": "{{ target_index_bulk_index_data_set_format }}",
+                "data_set_path": "{{ target_index_bulk_index_data_set_path }}"
+            },
+            "clients": {{ target_index_bulk_index_clients }}
+        },
+        {
+            "operation": {
+                "name": "refresh-target-index",
+                "operation-type": "custom-refresh",
+                "index": "{{ target_index_name }}",
+                "retries": 100
+            }
+        }
+    ]
+}
