+++ benchmarks/osb/procedures/train-test.json
@@ -1,28 +1,17 @@
 {% import "benchmark.helpers" as benchmark with context %}
 {
-    "version": 2,
-    "description": "k-NN Plugin train workload",
-    "indices": [
-        {
-            "name": "{{ target_index_name | default("target_index") }}",
-            "body": "target-index.json"
-        },
-        {
-            "name": "{{ train_index_name | default("train_index") }}",
-            "body": "train-index.json"
-        }
-    ],
-    "operations": {{ benchmark.collect(parts="operations/*.json") }},
+    "name": "train-test",
+    "default": false,
     "schedule": [
         {
-            "name": "delete-target-index",  
-            "index": "{{ target_index_name | default("target_index") }}",
+            "name": "delete-target-index",
+            "index": "{{ target_index_name }}",
             "only-if-exists": true,
             "operation": "delete-index"
         },
         {
             "name": "delete-train-index",
-            "index": "{{ train_index_name | default("train_index") }}",
+            "index": "{{ train_index_name }}",
             "only-if-exists": true,
             "operation": "delete-index"
         },
@@ -30,14 +19,14 @@
             "operation": {
                 "operation-type": "delete-model",
                 "name": "delete-model",
-                "model_id": "{{ model_id | default("test-model") }}"
+                "model_id": "{{ train_model_id }}"
             }
         },
         {
             "operation": {
                 "name": "create-train-index",
                 "operation-type": "create-index",
-                "index": "{{ train_index_name | default("train_index") }}"
+                "index": "{{ train_index_name }}"
             }
         },
         {
@@ -52,31 +41,31 @@
                 "name": "train-vector-bulk",
                 "operation-type": "custom-vector-bulk",
                 "param-source": "bulk-from-data-set",
-                "index": "{{ train_index_name | default("train_index") }}",
-                "field": "{{ train_field_name | default("train_field") }}",
-                "bulk_size": {{ train_bulk_size | default(200) }},
+                "index": "{{ train_index_name }}",
+                "field": "{{ train_field_name }}",
+                "bulk_size": {{ train_index_bulk_size }},
                 "data_set_format": "{{ train_index_data_set_format }}",
                 "data_set_path": "{{ train_index_data_set_path }}",
-                "num_vectors": {{ train_vector_limit | default(100000000000000000) }}
+                "num_vectors": {{ train_index_num_vectors }}
             },
-            "clients": {{ train_bulk_index_clients | default(1) }}
+            "clients": {{ train_index_bulk_index_clients }}
         },
         {
             "operation": {
                 "name": "refresh-train-index",
                 "operation-type": "custom-refresh",
-                "index": "{{ train_index_name | default("train_index") }}",
+                "index": "{{ train_index_name }}",
                 "retries": 100
             }
         },
         {
-            "operation": "{{ model_method }}-train-model"
+            "operation": "{{ train_model_method }}-train-model"
         },
         {
             "operation": {
                 "name": "create-target-index",
                 "operation-type": "create-index",
-                "index": "{{ target_index_name | default("target_index") }}"
+                "index": "{{ target_index_name }}"
             }
         },
         {
@@ -91,19 +80,19 @@
                 "name": "custom-vector-bulk",
                 "operation-type": "custom-vector-bulk",
                 "param-source": "bulk-from-data-set",
-                "index": "{{ target_index_name | default("target_index") }}",
-                "field": "{{ target_field_name | default("target_field") }}",
-                "bulk_size": {{ bulk_size | default(200) }},
-                "data_set_format": "{{ bulk_index_data_set_format }}",
-                "data_set_path": "{{ bulk_index_data_set_path }}"
-            },   
-            "clients": {{ bulk_index_clients | default(1) }}
+                "index": "{{ target_index_name }}",
+                "field": "{{ target_field_name }}",
+                "bulk_size": {{ target_index_bulk_size }},
+                "data_set_format": "{{ target_index_bulk_index_data_set_format }}",
+                "data_set_path": "{{ target_index_bulk_index_data_set_path }}"
+            },
+            "clients": {{ target_index_bulk_index_clients }}
         },
         {
             "operation": {
                 "name": "refresh-target-index",
                 "operation-type": "custom-refresh",
-                "index": "{{ target_index_name|default("target_index") }}",
+                "index": "{{ target_index_name }}",
                 "retries": 100
             }
         }
