+++ benchmarks/osb/workload.json
@@ -1,0 +1,17 @@
+{% import "benchmark.helpers" as benchmark with context %}
+{
+    "version": 2,
+    "description": "k-NN Plugin train workload",
+    "indices": [
+        {
+            "name": "{{ target_index_name }}",
+            "body": "{{ target_index_body }}"
+        },
+        {
+            "name": "{{ train_index_name }}",
+            "body": "{{ train_index_body }}"
+        }
+    ],
+    "operations": {{ benchmark.collect(parts="operations/*.json") }},
+    "test_procedures": [{{ benchmark.collect(parts="procedures/*.json") }}]
+}
